Event -> Removed the Month enum and started using java's own. Changed the date and time to LocalDate and LocalTime instead of String and int.Added an object constructor. Added getters and some setters, a function to output the event to a string and a method that allows the user to change desired values. Calendar -> Created functions for adding/removing Events to the calendar, checking for overlaps in the schedule, as well as multiple search and print functions. FileController -> Modified the read and write functions accordingly for the class changes. Added a merge function allowing the merging of data from a file with the loaded calendar. UI -> Made sure the UI is CLI and improved all menues. Finished the help function. Example.txt -> A couple of example test values.

+++ src/FileController.java
@@ -1,3 +1,6 @@
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.time.Month;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
@@ -10,15 +13,19 @@ public class FileController {
 
     Calendar calendar = Calendar.getInstance();
 
-    public FileController(){}
+    public FileController() {
+    }
 
-    public void readFile()
-    {
+    public void readFile() {
+        calendar.events = readFileToArray();
+    }
+
+    public List<Event> readFileToArray() {
         try {
             System.out.println("Enter the name of the text file you want to read from (without .txt at the end and without any dots)");
             Scanner scanner = new Scanner(System.in);
             String name = scanner.nextLine();
-            while(name.indexOf('.') != -1){
+            while (name.indexOf('.') != -1) {
                 System.out.println("No dots/file types allowed in the name!");
                 name = scanner.nextLine();
             }
@@ -27,92 +34,99 @@ public class FileController {
             Scanner myReader = new Scanner(myObj);
             List<Event> events = new ArrayList<>();
             while (myReader.hasNextLine()) {
-                Event e = new Event("", Month.JAN, 0, 0, 0, "");
+                Event e = new Event("", Month.JANUARY, 1, 0, 0, "");
 
                 String line = myReader.nextLine();
                 String[] data = line.split("-");
                 e.name = data[0];
                 e.desc = data[2];
                 data = data[1].split(" ");
-                e.month = Month.valueOf(data[0]);
-                e.day = Integer.parseInt(data[1]);
-                e.start = Integer.parseInt(data[2]);
-                e.end = Integer.parseInt(data[3]);
+                int year = Integer.parseInt(data[0]); //year
+                Month month = Month.valueOf(data[1]); // month
+                int day = Integer.parseInt(data[2]); //day
+
+                e.chosenDay = LocalDate.of(year, month, day);
+
+                e.startTime = LocalTime.parse(data[3]);
+                e.endTime = LocalTime.parse(data[4]);
 
                 events.add(e);
             }
             myReader.close();
-            if(events.isEmpty()) {
+            if (events.isEmpty()) {
                 System.out.println("File was empty, calendar has not been changed!");
-                UI ui = new UI();
-                ui.mainMenu();
             }
-            calendar.setEvents(events);
             System.out.println("Successfully read from the file.");
-            UI ui = new UI();
-            ui.mainMenu();
+            return events;
         } catch (FileNotFoundException e) {
             System.out.println("File doesn't exist!");
-            UI ui = new UI();
-            ui.mainMenu();
         }
+        return null;
     }
 
-    public void writeToTxt()
-    {
+    public void writeToTxt() {
+        if (calendar.events.isEmpty()) {
+            Scanner scanner = new Scanner(System.in);
+            System.out.println("You are about to write an empty calendar to a file, are you sure you want to continue?" +
+                    "\n Write 'CONTINUE' to go on. Write anything else to go back.");
+            String answer = scanner.nextLine();
+            if (!answer.equalsIgnoreCase("continue"))
+                return;
+        }
         try {
             System.out.println("Enter the name of the text file you want to write to (without .txt at the end and without any dots)");
             Scanner scanner = new Scanner(System.in);
             String path = scanner.nextLine();
-            while(path.indexOf('.') != -1){
+            while (path.indexOf('.') != -1) {
                 System.out.println("No dots/file types allowed in the name!");
                 path = scanner.nextLine();
             }
 
-            File wFile = new File(path +".txt");
+            File wFile = new File(path + ".txt");
             FileWriter myWriter = new FileWriter(wFile);
 
             //each event will be written on 1 row, every next element is spaced out, there's a - after the name and before the descriprion
-            for(Event e : calendar.events)
-            {
+            for (Event e : calendar.events) {
                 myWriter.write(e.name + "-");
-                myWriter.write(e.month + " ");
-                myWriter.write(e.day + " ");
-                myWriter.write(e.start + " ");
-                myWriter.write(e.end + "-");
+                myWriter.write(e.getYear() + " ");
+                myWriter.write(e.getMonth() + " ");
+                myWriter.write(e.getDay() + " ");
+                myWriter.write(e.startTime + " ");
+                myWriter.write(e.endTime + "-");
                 myWriter.write(e.desc);
                 myWriter.write("\n");
             }
 
             myWriter.close();
             System.out.println("Successfully wrote to the file.");
-            UI ui = new UI();
-            ui.mainMenu();
 
         } catch (IOException e) {
             System.out.println("An error occurred.");
             System.out.println(e.getMessage());
-            UI ui = new UI();
-            ui.mainMenu();
         }
     }
 
-    public void writeToFile()
-    {
+    public void writeToFile() {
+        if (calendar.events.isEmpty()) {
+            Scanner scanner = new Scanner(System.in);
+            System.out.println("You are about to write an empty calendar to a file, are you sure you want to continue?" +
+                    "\n Write 'CONTINUE' to go on. Write anything else to go back.");
+            String answer = scanner.nextLine();
+            if (!answer.equalsIgnoreCase("CONTINUE"))
+                return;
+        }
         try {
             System.out.println("Enter the name of the text file you want to write to, include the file type(Example -> .txt | .dat)[File name cannot contain 0]");
             System.out.println("WARNING: This function will create a file with whatever extension you choose for it, meaning the file you create might not be compatible with the information!");
             System.out.println("If you wish to return to the main menu at any time, please type '0' ");
             Scanner scanner = new Scanner(System.in);
             String path = scanner.nextLine();
-            if(path.equals("0"))
-            {
+            if (path.equals("0")) {
                 UI ui = new UI();
                 ui.mainMenu();
             }
-            while(path.indexOf('.') == -1){
-                if(path.equals("0"))
-                {
+            while (path.indexOf('.') == -1) {
+                if (path.equals("0")) {
                     UI ui = new UI();
                     ui.mainMenu();
                 }
@@ -125,29 +139,92 @@ public class FileController {
             FileWriter myWriter = new FileWriter(wFile);
 
             //each event will be written on 1 row, every next element is spaced out, there's a - after the name and before the descriprion
-            for(Event e : calendar.events)
-            {
+            for (Event e : calendar.events) {
                 myWriter.write(e.name + "-");
-                myWriter.write(e.month + " ");
-                myWriter.write(e.day + " ");
-                myWriter.write(e.start + " ");
-                myWriter.write(e.end + "-");
+                myWriter.write(e.getYear() + " ");
+                myWriter.write(e.getMonth() + " ");
+                myWriter.write(e.getDay() + " ");
+                myWriter.write(e.startTime + " ");
+                myWriter.write(e.endTime + "-");
                 myWriter.write(e.desc);
                 myWriter.write("\n");
             }
 
             myWriter.close();
             System.out.println("Successfully wrote to the file.");
-            UI ui = new UI();
-            ui.mainMenu();
 
         } catch (IOException e) {
             System.out.println("An error occurred.");
             System.out.println(e.getMessage());
-            UI ui = new UI();
-            ui.mainMenu();
         }
     }
 
-    public void mergeData(){}
+    public void mergeData() {
+        List<Event> newEvents = readFileToArray();
+
+        for (Event e : newEvents) {
+            System.out.println("Next event: \n");
+            System.out.println(e.ShowEvent());
+            System.out.println("Trying to add next event... \n");
+            List<Event> overlaps = calendar.findOverlap(e);
+            if (!overlaps.isEmpty()) {
+                boolean repeat = false;
+                System.out.println("Old event/s:\n");
+                for (Event ev : overlaps)
+                    System.out.println(ev.ShowEvent());
+
+                System.out.println("New event:\n");
+                System.out.println(e.ShowEvent());
+
+                do {
+                    System.out.println("Found overlapping event/s!\n");
+                    System.out.println("Please select the next action:\n");
+                    System.out.println("0: Keep the old event\n");
+                    System.out.println("1: Keep the new event\n");
+                    System.out.println("2: Change the old event\n");
+                    System.out.println("3: Change the new event\n");
+                    System.out.println("-1: Return to menu\n");
+
+                    Scanner scanner = new Scanner(System.in);
+                    while (!scanner.hasNextInt()) {
+                        System.out.println("Error, invalid number entered!");
+                        scanner.nextLine();
+                    }
+
+                    int action = scanner.nextInt();
+
+                    switch (action) {
+                        case 0: //Move to next event
+                            continue;
+                        case 1: //Replace the old event/s
+                            for (Event ev : overlaps)
+                                calendar.removeEvent(ev);
+                            calendar.book(e);
+                            break;
+                        case 2: //Move the old event/s
+                            for (Event ev : overlaps)
+                                calendar.change(ev);
+                            repeat = !calendar.book(e);
+                            if (repeat)
+                                System.out.println("You did not clear the schedule around the new event!");
+                            break;
+                        case 3://Move the new event
+                            e.change();
+                            repeat = !calendar.book(e);
+                            if (repeat)
+                                System.out.println("You did not choose a free time for the new event!");
+                            break;
+                        case -1:
+                            return;
+                        default:
+                            System.out.println("Number out of range! Please choose a valid option.");
+                    }
+                } while (repeat);
+            } else {
+                calendar.book(e);
+                System.out.println("Successfully added!\n");
+            }
+        }
+
+    }
 }
